requirements.txt
Django==2.2.17
djangorestframework==3.12.2
geographiclib==1.50
geopy==2.1.0
pytz==2020.5
sqlparse==0.4.1
